DB Serialize to Types (#182)
* mod

* mod

* prepare

+++ backend/crates/repository/src/pg/project.rs
@@ -24,7 +24,7 @@ fn create_project<'a, 'c, Connection: Acquire<'c, Database = Postgres> + Send + 
 
         let project = sqlx::query_as!(
             Project,
-            r#"INSERT INTO projects (tenant, id, fhir_version, system_created) VALUES ($1, $2, $3, $4) RETURNING tenant, id, fhir_version as "fhir_version: SupportedFHIRVersions""#,
+            r#"INSERT INTO projects (tenant, id, fhir_version, system_created) VALUES ($1, $2, $3, $4) RETURNING tenant as "tenant: TenantId", id as "id: ProjectId", fhir_version as "fhir_version: SupportedFHIRVersions""#,
             tenant.as_ref(),
             id.as_ref(),
             project.fhir_version as SupportedFHIRVersions,
@@ -47,7 +47,7 @@ fn read_project<'a, 'c, Connection: Acquire<'c, Database = Postgres> + Send + 'a
         let mut conn = connection.acquire().await.map_err(StoreError::SQLXError)?;
         let project = sqlx::query_as!(
             Project,
-            r#"SELECT id, tenant, fhir_version as "fhir_version: SupportedFHIRVersions" FROM projects where tenant = $1 AND id = $2"#,
+            r#"SELECT id as "id: ProjectId", tenant as "tenant: TenantId", fhir_version as "fhir_version: SupportedFHIRVersions" FROM projects where tenant = $1 AND id = $2"#,
             tenant.as_ref(),    
             id
         )
@@ -68,7 +68,7 @@ fn delete_project<'a, 'c, Connection: Acquire<'c, Database = Postgres> + Send + 
         let mut conn = connection.acquire().await.map_err(StoreError::SQLXError)?;
         let deleted_project = sqlx::query_as!(
             Project,
-            r#"DELETE FROM projects WHERE tenant = $1 AND id = $2 and system_created = false RETURNING id, tenant, fhir_version as "fhir_version: SupportedFHIRVersions""#,
+            r#"DELETE FROM projects WHERE tenant = $1 AND id = $2 and system_created = false RETURNING id as "id: ProjectId", tenant as "tenant: TenantId", fhir_version as "fhir_version: SupportedFHIRVersions""#,
             tenant.as_ref(),
             id
         )
